#8 게시글 삭제 API 수정 - 토큰을 검사하여, 유효한 토큰이면서 해당 사용자가 작성한 게시글만 삭제 가능

+++ src/main/java/com/sparta/springlv2/service/BoardService.java
@@ -46,8 +46,11 @@ public class BoardService {
         return ResponseEntity.ok().body(new BoardResponseDto.BoardReadResponseDto(board)).getBody();
     }
 
-    public boolean deleteBoard(Long id, BoardRequestDto requestDto){
+    public boolean deleteBoard(Long id, BoardRequestDto requestDto, User user){
         Board board = findBoard(id);
+        if(!Objects.equals(board.getUser().getId(), user.getId())){
+            throw new IllegalArgumentException("해당 사용자가 작성한 게시글이 아닙니다.");
+        }
         boardRepository.delete(board);
         return true;
     }
